LALInferenceReadData correcte SNR srqt(2) factor off, and LALInspiralTemplate fixed unsigned int substractions Original: 78eb8093cdf8fd7f182f323aad7b374019cd59b2

+++ lalinference/src/LALInferenceTemplate.c
@@ -1807,10 +1807,10 @@ void LALInferenceTemplateLALGenerateInspiral(LALInferenceIFOData *IFOdata)
                 IFOdata->timeModelhPlus->data->data[i] = 0;
                 IFOdata->timeModelhCross->data->data[i] = 0;		
               }else{
-                a1		= waveform.a->data->data[2*(i-windowshift)];
-                a2		= waveform.a->data->data[2*(i-windowshift)+1];
-                phi     = waveform.phi->data->data[i-windowshift];
-                if (waveform.shift) shift   = waveform.shift->data->data[i-windowshift];
+                a1		= waveform.a->data->data[2*((INT8)i-(INT8)windowshift)];
+                a2		= waveform.a->data->data[2*((INT8)i-(INT8)windowshift)+1];
+                phi     = waveform.phi->data->data[(INT8)i-(INT8)windowshift];
+                if (waveform.shift) shift   = waveform.shift->data->data[(INT8)i-(INT8)windowshift];
                 else shift = 0.0;
 					
                 IFOdata->timeModelhPlus->data->data[i] = a1*cos(shift)*cos(phi) - a2*sin(shift)*sin(phi);
@@ -1824,14 +1824,14 @@ void LALInferenceTemplateLALGenerateInspiral(LALInferenceIFOData *IFOdata)
               fprintf(stderr, "The waveform template used will be missing its first %d points. Consider increasing the segment length (--seglen). (in %s, line %d)\n",waveform.phi->data->length - IFOdata->timeData->data->length + (int) windowshift , __FILE__, __LINE__);
             }
             for (i=0; i<IFOdata->timeData->data->length; i++){
-              if(i>=IFOdata->timeData->data->length-windowshift){
+              if((INT8)i>=(INT8)IFOdata->timeData->data->length-(INT8)windowshift){
                 IFOdata->timeModelhPlus->data->data[i] = 0.0;
                 IFOdata->timeModelhCross->data->data[i] = 0.0;
               }else{
-                a1		= waveform.a->data->data[2*(i+waveform.phi->data->length-IFOdata->timeData->data->length+windowshift)];
-                a2		= waveform.a->data->data[2*(i+waveform.phi->data->length-IFOdata->timeData->data->length+windowshift)+1];
-                phi     = waveform.phi->data->data[i+waveform.phi->data->length-IFOdata->timeData->data->length+windowshift];
-                if (waveform.shift) shift   = waveform.shift->data->data[i+waveform.phi->data->length-IFOdata->timeData->data->length+windowshift];
+                a1		= waveform.a->data->data[2*((INT8)i+(INT8)waveform.phi->data->length-(INT8)IFOdata->timeData->data->length+(INT8)windowshift)];
+                a2		= waveform.a->data->data[2*((INT8)i+(INT8)waveform.phi->data->length-(INT8)IFOdata->timeData->data->length+(INT8)windowshift)+1];
+                phi     = waveform.phi->data->data[(INT8)i+(INT8)waveform.phi->data->length-(INT8)IFOdata->timeData->data->length+(INT8)windowshift];
+                if (waveform.shift) shift   = waveform.shift->data->data[(INT8)i+(INT8)waveform.phi->data->length-(INT8)IFOdata->timeData->data->length+(INT8)windowshift];
                 else shift = 0.0;
               
                 IFOdata->timeModelhPlus->data->data[i] = a1*cos(shift)*cos(phi) - a2*sin(shift)*sin(phi);
@@ -1858,8 +1858,8 @@ void LALInferenceTemplateLALGenerateInspiral(LALInferenceIFOData *IFOdata)
                 IFOdata->timeModelhPlus->data->data[i] = 0;
                 IFOdata->timeModelhCross->data->data[i] = 0;		
               }else{
-                IFOdata->timeModelhPlus->data->data[i] = waveform.h->data->data[2*(i-windowshift)];
-                IFOdata->timeModelhCross->data->data[i] = waveform.h->data->data[2*(i-windowshift)+1];
+                IFOdata->timeModelhPlus->data->data[i] = waveform.h->data->data[2*(i-(INT8)windowshift)];
+                IFOdata->timeModelhCross->data->data[i] = waveform.h->data->data[2*(i-(INT8)windowshift)+1];
               }
             }
           }else{
@@ -1869,12 +1869,12 @@ void LALInferenceTemplateLALGenerateInspiral(LALInferenceIFOData *IFOdata)
               fprintf(stderr, "The waveform template used will be missing its first %d points. Consider increasing the segment length (--seglen). (in %s, line %d)\n",waveform.h->data->length - IFOdata->timeData->data->length + (int) windowshift , __FILE__, __LINE__);
             }
             for (i=0; i<IFOdata->timeData->data->length; i++){
-              if(i>=IFOdata->timeData->data->length-windowshift){
+              if((INT8)i>=(INT8)IFOdata->timeData->data->length-(INT8)windowshift){
                 IFOdata->timeModelhPlus->data->data[i] = 0.0;
                 IFOdata->timeModelhCross->data->data[i] = 0.0;
               }else{                
-                IFOdata->timeModelhPlus->data->data[i] = waveform.h->data->data[2*(i+waveform.h->data->length-IFOdata->timeData->data->length+windowshift)];
-                IFOdata->timeModelhCross->data->data[i] = waveform.h->data->data[2*(i+waveform.h->data->length-IFOdata->timeData->data->length+windowshift)+1];
+                IFOdata->timeModelhPlus->data->data[i] = waveform.h->data->data[2*((INT8)i+(INT8)waveform.h->data->length-(INT8)IFOdata->timeData->data->length+(INT8)windowshift)];
+                IFOdata->timeModelhCross->data->data[i] = waveform.h->data->data[2*((INT8)i+(INT8)waveform.h->data->length-(INT8)IFOdata->timeData->data->length+(INT8)windowshift)+1];
               }
             }
           instant-= ((INT8)waveform.h->data->length-(INT8)IFOdata->timeData->data->length+2*(INT8)windowshift)*IFOdata->timeData->deltaT;
